#2074 ChatService для создания и редактирования сообщений использует хэндлеры; Хэндлеры возвращают соответствующие экземпляры MessageBroadcast;

+++ wfe-core/src/main/java/ru/runa/wfe/chat/logic/ChatLogic.java
@@ -119,7 +119,9 @@ public class ChatLogic extends WfCommonLogic {
         ChatMessage message = getMessageById(user, messageId);
         List<ChatMessageFile> files = fileLogic.getByMessage(user, message);
         messageDao.deleteMessage(messageId);
-        fileLogic.delete(user, files);
+        if (files != null) {
+            fileLogic.delete(user, files);
+        }
     }
 
     public void updateMessage(User user, ChatMessage message) {
